[#2] Extract user creation to the auth service
