+++ nextjs/lib/error.ts
@@ -1,0 +1,16 @@
+export class BaseError extends Error {
+  constructor({
+    message,
+  }: {
+    message: string
+  }) {
+    // Pass remaining arguments (including vendor specific ones) to parent constructor
+    super(message)
+
+    if (Error.captureStackTrace) {
+      Error.captureStackTrace(this, BaseError)
+    }
+  }
+}
+
+export default BaseError
