+++ nextjs/package.json
@@ -12,7 +12,6 @@
     "@prisma/client": "4.3.1",
     "@types/passport": "1.0.11",
     "@types/passport-google-oauth20": "2.0.11",
-    "axios": "0.27.2",
     "iron-session": "6.2.1",
     "next": "12.3.0",
     "next-connect": "0.13.0",
