+++ nextjs/services/auth/error.ts
@@ -1,0 +1,3 @@
+import BaseError from '../../lib/error'
+
+export default class AuthError extends BaseError {}
