+++ nextjs/services/auth/google.service.ts
@@ -1,5 +1,9 @@
-import { passport } from '../../lib/middleware/passport.middleware'
+import { User } from '@prisma/client'
+import { Profile } from 'passport'
 
+import { passport } from '../../lib/middleware/passport.middleware'
+import * as UserRepository from '../../repositories/user.repository'
+import AuthError from './error'
 class AuthGoogleService {
   provider: any
 
@@ -8,7 +12,29 @@ class AuthGoogleService {
   }
 
   authenticate() {
-    return this.provider.authenticate('google')
+    try {
+      return this.provider.authenticate('google')
+    } catch (error: any) {
+      throw new AuthError(error)
+    }
+  }
+
+  static async verifyOrCreateUser(userProfile: Profile): Promise<User | AuthError> {
+    try {
+      const { photos, displayName, emails } = userProfile
+      const existingUser = await UserRepository.findUserByEmail(emails[0].value)
+      
+      if (existingUser) {
+        return Promise.resolve(existingUser)
+      } 
+
+      const userAttributes = { avatarUrl: photos[0].value, name: displayName, email: emails[0].value }
+      const newUser = await UserRepository.createUser(userAttributes)
+
+      return Promise.resolve(newUser);
+    } catch (error: any) {
+      return Promise.reject(new AuthError({message: 'User could not be verified or created' }));
+    }
   }
 }
